+++ zebra-rpc/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `methods::types::get_block_template::fetch_chain_info`
+
 ## [3.1.0] - 2025-11-17
 
 ## Added
@@ -14,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Republicized `valid_addresses` method ([#10021](https://github.com/ZcashFoundation/zebra/pull/10021))
-
 
 ## [3.0.0] - 2025-10-15